Refunctioned to a diary-style app. All reviews have creators, and the main container will only display reviews uploaded by the user currently logged in.

+++ project-reviews/src/Components/ReviewsContainer/ReviewsList/ReviewsList.jsx
@@ -3,19 +3,23 @@ import EditReview from '../EditReviewModal/EditReviewModal'
 
 const Reviews = (props) =>{
     const reviewList = props.reviews.map((review, i)=>{
-        return(
-            <li key={review._id}>
-                <h5>{review.title}</h5><br/>
-                <p>{review.description}</p>
-                <small>{review.review}</small>
-                <small>{review.rating}</small>               
-                <EditReview review={review} updateReview={props.updateReview}/>
-                <button onClick={(e)=>{
-                    e.preventDefault();
-                    props.deleteReview(review._id)
-                }}>Delete</button>
-            </li>
-        )
+          if(review.creator===props.userID){
+            console.log(review)
+            return(
+                <li key={review._id}>
+                    <h5>{review.title}</h5><br/>
+                    <p>{review.description}</p>
+                    <p>{review.playthrough}</p>
+                    <small>{review.review}</small><br/>
+                    <small>{review.rating}</small>               
+                    <EditReview review={review} updateReview={props.updateReview}/>
+                    <button onClick={(e)=>{
+                        e.preventDefault();
+                        props.deleteReview(review._id)
+                    }}>Delete</button>
+                </li>
+            )
+         }
     })
     return(
         <ul>{reviewList}</ul>
